dbutil maxOpenConns
