+++ src/internal/storage/track/renewer.go
@@ -60,13 +60,14 @@ func (r *Renewer) Add(ctx context.Context, id string) error {
 
 // Close stops the background renewal
 func (r *Renewer) Close() (retErr error) {
+	ctx := context.Background()
 	if _, err := r.tracker.SetTTLPrefix(ctx, r.id+"/", ExpireNow); retErr == nil {
 		retErr = err
 	}
 	if err := r.r.Close(); retErr == nil {
-		return
+		retErr = err
 	}
-	return r.r.Close()
+	return retErr
 }
 
 func (r *Renewer) nextInt() int {
